requirements.txt
python==3.6
numpy
conda
rdkit-pypi
gunicorn
Flask
Jinja2
Werkzeug
distribute
wsgiref